reordered folders and added unit tests for visualization
