+++ README.md
@@ -1,1 +1,2 @@
+![develop status](https://github.com/duyanhpham-brs/XAI-Multivariate-Time-Series/workflows/develop_branch/badge.svg)
 # XAI-Multivariate-Time-Series
