+++ environment.yml
@@ -25,6 +25,7 @@ dependencies:
   - icu=58.2=h0a44026_3
   - importlib-metadata=3.7.3=py37hecd8cb5_1
   - importlib_metadata=3.7.3=hd3eb1b0_1
+  - iniconfig=1.1.1=pyhd3eb1b0_0
   - intel-openmp=2019.4=233
   - ipykernel=5.3.4=py37h5ca1d4c_0
   - ipython=7.21.0=py37h01d92e1_0
@@ -65,6 +66,7 @@ dependencies:
   - mkl-service=2.3.0=py37h9ed2024_0
   - mkl_fft=1.3.0=py37ha059aab_0
   - mkl_random=1.1.1=py37h959d312_0
+  - more-itertools=8.7.0=pyhd3eb1b0_0
   - nbclient=0.5.3=pyhd3eb1b0_0
   - nbconvert=6.0.7=py37_0
   - nbformat=5.1.2=pyhd3eb1b0_1
@@ -75,7 +77,7 @@ dependencies:
   - numpy=1.19.2=py37h456fd55_0
   - numpy-base=1.19.2=py37hcfb5961_0
   - olefile=0.46=py37_0
-  - openssl=1.1.1j=h9ed2024_0
+  - openssl=1.1.1k=h9ed2024_0
   - packaging=20.9=pyhd3eb1b0_0
   - pandas=1.2.3=py37hb2f4e1b_0
   - pandoc=2.12=hecd8cb5_0
@@ -86,15 +88,18 @@ dependencies:
   - pickleshare=0.7.5=pyhd3eb1b0_1003
   - pillow=8.1.2=py37h5270095_0
   - pip=21.0.1=py37hecd8cb5_0
+  - pluggy=0.13.1=py37hecd8cb5_0
   - prometheus_client=0.9.0=pyhd3eb1b0_0
   - prompt-toolkit=3.0.8=py_0
   - prompt_toolkit=3.0.8=0
   - ptyprocess=0.7.0=pyhd3eb1b0_2
+  - py=1.10.0=pyhd3eb1b0_0
   - pycparser=2.20=py_2
   - pygments=2.8.1=pyhd3eb1b0_0
   - pyparsing=2.4.7=pyhd3eb1b0_0
   - pyqt=5.9.2=py37h655552a_2
   - pyrsistent=0.17.3=py37haf1e3a3_0
+  - pytest=6.2.2=py37hecd8cb5_2
   - python=3.7.10=h88f2d9e_0
   - python-dateutil=2.8.1=pyhd3eb1b0_0
   - pytorch=1.7.1=py3.7_0
@@ -116,6 +121,7 @@ dependencies:
   - testpath=0.4.4=pyhd3eb1b0_0
   - threadpoolctl=2.1.0=pyh5ca1d4c_0
   - tk=8.6.10=hb0a8c7a_0
+  - toml=0.10.2=pyhd3eb1b0_0
   - torchaudio=0.7.2=py37
   - torchvision=0.8.2=py37_cpu
   - tornado=6.1=py37h9ed2024_0
