+++ feature_extraction/CAMs/ActivationSmoothScoreCAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.ActivationSmoothScoreCAM.ActivationSmoothScoreCAM import ActivationSmoothScoreCAM
