+++ feature_extraction/CAMs/CAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.CAM.CAM import CAM
