+++ feature_extraction/CAMs/GradCAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.GradCAM.GradCAM import GradCAM
