+++ feature_extraction/CAMs/GradCAMPlusPlus/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.GradCAMPlusPlus.GradCAMPlusPlus import GradCAMPlusPlus
