+++ feature_extraction/CAMs/InputSmoothScoreCAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.InputSmoothScoreCAM.InputSmoothScoreCAM import InputSmoothScoreCAM
