+++ feature_extraction/CAMs/IntegratedScoreCAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.IntegratedScoreCAM.IntegratedScoreCAM import IntegratedScoreCAM
