+++ feature_extraction/CAMs/ScoreCAM/__init__.py
@@ -1,0 +1,1 @@
+from feature_extraction.CAMs.ScoreCAM.ScoreCAM import ScoreCAM
